memmap-storage: Allow instructions to create an i2c device on sysfs
Follow same 'create' format of `sol_iio_address_device`. Diferently,
however, it's a synchronous operation, so no changes to `sol_memmap_add`
interface.
To achieve this, sol-fbp-generator and sol-conffile were adapted to new
sol_memmap_map structure, that has now a pointer to an array of entries,
instead of array of entries directly.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>