comms/oic-client: Implement discovery timeout to avoid race condition
If a CoAP packet is confirmable, the CoAP layer won't call user callbacks
once a response has been received.  That doesn't happen with non-confirmable
callbacks, which is the case for discovery packets that are broadcast to the
network.

Add pending discovery contexts to a list, and remove them after 10s. Only
then free them.  Discard response packet if not in that list.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>