util: Ensure align_power2() works with things other than uint
This creates a bunch of align_power2() functions to work with ingegers
of different sizes.  C11's _Generic() would be a better candidate but
for some reason it's not available in the CI server.

These functions will each return the correct maximum value and uses the
correct leading zero counting builtin for the type, and, in the case of
unsigned shorts (often 16-bit), provide a function that clamps the
results to USHRT_MAX.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>