Add 'sol' tool
A client/server software writen in Go to make easy building things
remotely. The idea is that the server side could be in a VM or a server
under user control, and the client side is the only code that in the
user's environment.

Motivation for write this in Go was: the standard library contained all
the tools we needed (no external libraries), the single binary for the
client code could be easily cross-compiled to run on Windows or Mac.

Refer to README to learn how to use it.

Signed-off-by: Luiz Ywata <luizg.ywata@intel.com>