sol_buffer: Fix append_printf function bug
Function sol_buffer_append_printf was not adding the last character to
the buffer if SOL_BUFFER_FLAGS_NO_NUL_BYTE was used and if this
character would be placed in the last available buffer position.

This was happening because printf function always place '\0' at the end
of the string being printed. So we need to allocate an extra byte even
if SOL_BUFFER_FLAGS_NO_NUL_BYTE is used.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>