flow: add new selector node, with tests.
Quoting the Kconfig help verbatim:

"It is common to have IoT products with LCD display and buttons (arrows)
and the need to ask for user input. This is a collection of high-level
node types to deal with that, the *selector* nodes.

These node types format a multi-line string to be used by other node
types doing visual output, such as LCD display (grove/lcd-string), in
conjuntion with others providing input."

I did my best to have broad tests, but much more than that was tested by
hand while trying to make the code robust. It's a PITA to write
validation to this node, so let's grow that with the time.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>