comms:flow:http: Add node to serve static files
This node has an input port do enable/disable (similar the timer's
node). The path that contains the files is set trough the option 'path'
and the there is one more option to set if the files should be served
automatically on the node's open, by default this option is true.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>