comms: Send multicast packets to all interfaces
On multi-homed systems, OIC discovery messages were being sent to only
one interface, often not the desired one.  Because of this, it would
fail to find devices.

Modify sol_socket::sendmsg() so that, whenever sending data to a
multicast address, it loops through all interfaces that are running and
multicast capable, calling sendmsg(2) with a `struct ip_pktinfo` (and
its IPv6 counterpart, `struct ip6_pktinfo`) as ancillary data with the
network interface index.  Consider a success if at least one of these
sendmsg(2) calls succeed.

This change is only required in the Linux implementation at the moment,
and is completely transparent on an API level.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>