comms: Send multicast packets to all interfaces
On multi-homed systems, OIC discovery messages were being sent to only
one interface, often not the desired one.  Because of this, it would
fail to find devices.

Modify sol_socket::sendmsg() so that, whenever sending data to a
multicast address, it loops through all interfaces that are running and
multicast capable, setting `IP_MULTICAST_IF` or `IPV6_MULTICAST_IF` (as
appropriate per the address family), sending the packet, and then
restoring the original option.

This change is only required in the Linux implementation at the moment,
and is completely transparent on an API level.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>