comms: Send multi packets to all interfaces
On multi-homed systems, OIC discovery messages was being sent to only
one interface, often not the desired one.  Because of this, it would
often fail to find devices.

Modify sol_socket::sendmsg() so that, whenever sending data to a
multicast address, it loops through all interfaces that are running and
multicast capable, and sets a `struct ip_pktinfo` (and its IPv6
counterpart, `struct ip6_pktinfo`) as ancillary data.  This allows
setting the network interface index that `sol_network` is already aware
of.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>