file-reader: introduce sol_file_reader_fd() API
This patch introduces the sol_file_reader_fd() API. This function is
similar to sol_file_reader_open() but instead of an file path argument
it expects a file descriptor.

This call doesn't actually open the file but expects its caller to do
so, having said that, the caller also must ensure it's properly closed
after its use.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>