lcd: unrotten grove/lcd* nodes.
These nodes had serious invalid memory access problems in its
asynchronous callbacks. This is so becase the ordered commands live in a
Soletta vector, which by design does reallocs on its contents from time
to time. We can optimise it later, but for now, the content passed to
the I2C calls is allocated from the original commands, to be freed after
use.

Various other minor fixes to the node were made together.

Finally, there was no reason for the grove/lcd-string node not to have
its color exposed as an option too -- now done.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>