Sol-fbp-generator: Generate missing option values for nodes.
When defining a rgb constant node for example, depending
how the fbp code is written some values of the sol_rgb struct
would not be set. This could could lead to wrong result
as some node expects that the all fields are initialized.

With this patch the option fields that are not set by the user,
will be set with its default values.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>