builders: sol -run-as-server fail if no valid platforms
A common mistake was to not run 'sol -run-as-server' inside the prepared
directory (builders/out/). This change makes sure we fail early, instead
of until compile happens.

The system uses the fact that compile must be executable for the
platform directory to be ready. We may in the future rename compile in
the "source" directories to avoid mistakes of confusing prepared
directories and source directories.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>