Generator failed when connecting 2 nodes using composed types
When a composed type port from a meta generated node is connected to a
composed type port from a custom node, generator fails to check if the
types are the same. This happens because json used one format to
describe composed packets and meta nodes uses other format.

With this patch, both cases will use the same format for representing
composed types

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>