memmap-storage: Add per map write timeout
Writings to a given file [EEPROM or NVRAM device] will be grouped to be
executed on a timeout defined via its map.
APIs to get and set timeout also added.
test-persistence-memmap updated to test this feature, specially to check
if multiple files are handled correctly.
To ease implementation, sol-memmap-storage now keeps all internal
information about a map (resolved path, timeout, if it's checked) on a
new struct `map_internal`.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>