Set packet type name to same name used when defining nodes
The packet types name property was different from the name used when
defining the type of a port in a node definition. This may be confusing
for users.

With this change it is also possible for us to use the packet type name
property in composed type handling, instead of needing to convert them
to the correct type name.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>