flow: change 'float' options to 'drange'
Since it's not a single float, it's a full drange data,
the name was misleading.

Actually, after some discussion, we get to conclusion this
type will be dropped, and instead another two should
be used:

    * float for single floats
    * float_spec (for min / max / step)

If drange is wanted, two options will be required (float +
drange_spec).
So, for now, a "deprecated" warning will be displayed when
a type using it is generated.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>