comms: Add generic MQTT flow module
This commit adds a generic client node for MQTT that can be used for
publishing messages and subscribing to topics.

The parameters for the MQTT connection can be configured both by config
and by sending a message to the specific flow port. The connection is
triggered when the first publish or subscribe trigger is activated and
is maintained afterwards. Connection problems are notified through the
ERROR out port.

Signed-off-by: Rodrigo Chiossi <rodrigo.chiossi@intel.com>