linux-micro: attach a gdbserver if on debug mode
This patch introduces the sol-debug=1 and sol-debug-comm=* arguments to
be used when linux-micro is running as pid1.

If these arguments are provided by kernel command line, the linux micro
platform implementation will exec() a gdbserver process and attach it to
the current pid1 - as soon as the platform init process takes place.

The sol-debug-comm=* argument is used to determine the comm to be used
by the gdbserver and if not provided /dev/ttyS0 is assumed as default.

Tested with kvm, adding a serial chardev, and using it as target remote.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>