persistence: Force sending packets on reset
Now that writes are async, the write callback when reseting
a value can take too long and be superseeded by a new value being
writen to the node. So, to be able to 'reset to default' when
having no previous stored value, persistence node sends a packet
right after reseting its value - it doesn't wait for storage write
callback anymore.
This is useful when opening a node that has no previous value: without
it, there would be no packet being sent on node opening.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>