HTTP: Replace "const char *" variables at sol_http_param_value to struct sol_str_slice.
This refactoring will be usefull in order to save memory when the "create URL",
APIs are ready. The reason for this, is that there will be an API
that given an URL, it will split all its components
(scheme, server, port, paths, params and etc). All these components
will be stored in sol_str_slices in order to prevent memory allocations and
will be returned to the user,
however HTTP params such as Headers, cookies, and query would have to be
duplicated as struct sol_http_param_value stores them as "const char *".

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>