build: fix Makefile.gen being dependent on kconfig's .config file.
The former is the base for pratically all the build, but .config should
only be kconfig's business (make *config output), which also may update
Makefile.gen.

With this change, editions to .config after kbuild invocation won't
trigger Makefile.gen (deps checking) again, facilitating the build for
packaging, where we make an alldefconfig call and edit with sed just the
bits that we want (-g for CFLAGS, even being a release build, etc).

In order for this to work, a circular dependency that were there was
removed -- some name/path definitions that were placed in the
dependencies.json file were move to the build infra Makefiles. The
dependency-resolver script used to generate a .h file for inclusion with
those definitions, that are now passed via CFLAGS to every object.