resolver: Improve warning about module not found
Because dlerror() prints the full path used in the erroneous attempt,
so the final warning message was awkwardly long and containing redundant
information.

This replaces the first path with the module name used to form the path
and move 'dlerror' message to a new line, indented to make sure that is
part of the previous line.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>