sol-oic-gen: add SCAN port on clients
With that it's possible to request a scan for all servers
that matches the same interface of clients.

IDs will be dispatched on DEVICE_ID port, so later
they can be used to select the device which the client
will talk.

Ideally a human-friendly name should be sent on packet
as well, not only the ID string, but I didn't found
anything like that in our implementation so far.

sol_oic_client_get_server_info return some information,
but only stuff related to a model, so users wouldn't
be able to use it to differentiate devices.

Maybe manufacturer_name + model_number could be used.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>