sol-oic-gen: add DEVICE_ID port on clients
As client nodes support scanning for servers that matches this client
interface, we needed a way of setting in a client one of the device
IDs scanned.

Option device_id is now optional. If set, the client node will try to
find a server with the specified ID. If not, node will wait a packet in
DEVICE_ID port before trying to connect to a specific server.

When a device ID is received in DEVICE_ID port, the client node will
stop to talk with the previous selected server and try to look for a new
server using the new device_id

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>