Flow: Remove int size check in http-client.c
This type of check shoud be make inside the http-client
implementation (in this case, out curl backend) not in the node.
However, we're only using the curl under Linux so the size check
is not necessary.
If a new backend is added, the overflow check should be made there.

In order to create a sane API, the timeout will now be an int32_t.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>