oic: Improve oic-client-scan demo
Improve oic-client-demo by adding a selector to choose which light
deivice will be controled by the client node. After choosing a device,
user is able to turn on/off the light.

Persistence is also used to save the favorite device and to auto-select
it when possible. Changing the device being controlled is also
supported.

Two json conffile are provided, one for a gtk backend demo and another
to run in an Edison board.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>