coap: reply from multicast packets was ignored after first reply
When sol_coap_send_packet_with_reply function was called sol-coap was
waiting for first reply and then ignoring everything else. This may be
ok when we are talking with a single server, but it is a problem when
using multicast packets to discovery a server.

The approach we took now is to look at a boolean return value of the
reply callback to decide if we can ignore following packets of if we
need to keep waiting a reply.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>