coap: stop waiting for a reply after timeout
When sol_coap_send_packet_with_reply is called but no reply is received
or reply_cb always returned true we won't remove pending_reply struct
and we'll keep waiting for replies forever.
When we failed to deliver a packet sent using
sol_coap_send_packet_with_reply and maximum number of retries is
reached, we stop retrying to send the packet, but don't remove
pending_reply struct too.

With this patch users of sol_coap_send_packet_with_reply will be warned
when a timeout occurs giving them a chance to stop waiting for new
replies.

Any ajustments in oic functions and API was performed to support these
changes.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>