Sol-HTTP-Client: Properly parse the URL given from options and ports.
Prior to this patch if one provide the following URL
"http://www.example.com/?ILoveHttpQueryParams=TheyAreAwesome" using the URL
port or the node option and later add more params using the port PARAM_QUERY,
the final URL would be:
"http://www.example.com/?ILoveHttpQueryParams=TheyAreAwesome?ThisUrlIsWrong=YES&Cries".

It's clear that the final URL is malformed. In order to prevent this bug,
when an URL is provided it will be properly splitted so this kind of
problem won't occur.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>