coap: Add api to cancel a packet scheduled to be send
Make it possible to cancel a packet that was scheduled to be sent using
sol_coap_send_packet or sol_coap_send_packet_with_reply.

If the packet was not sent yet, it will be canceled
If we tried to send the packet, but sending it failed, we won't try to
send it anymore.
In all cases, even if the packet was successfully sent we won't call
reply_cb when a reply arrives.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>