update: New module for Soletta - update. So Solleta apps ca be updated.
This commit introduces an API to update Soletta based apps, and
implements one module to show it working. The overall idea is that
different update modules can be used on different Soletta scenarios.
The API is quite straightforwad, with methods to check for updates,
fetch and install them.
The module provided in this commit is for scenario of Soletta app
statically built running as PID1, so update Soletta is essencially
changing app current file. New modules can be added on future, handling
dependencies or other scenarios.
A common set of functions also made available to ease update module
development, such as common check and fetch functions.
Currently, only hash/digest of file is verified, but it's very important
to add a way to check signatures, so files are from a trusted source.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>