riot: Refactor interrupt scheduler to avoid malloc() on interruptions
Instead of allocating memory to store the read byte from the serial,
use a buffer allocated when the interruptions are enabled. This serves
two purposes, avoids the allocations in interrupt context, which are
risky, and helps reduce the number of messages that will be queued on
the main thread for each byte.

Signed-off-by: Iván Briano <ivan.briano@intel.com>