Sol-flow-composed: Change sol_flow_packet_get_composed_members behavior
When this function was created it was supposed to just count the
composed members len and one could iterate over a list of sol_flow_packet.
However with the introduction of sol_flow_packet_get_composed_members()
this function became useless and no one was using it.

The function sol_flow_packet_get_composed_members() was refactored to return
which packet types one composed type is made of (it was also renamed).

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>