coap: Improve handing of cancelations of observing packets
When reply_cb returns false we are suppose to stop waiting for new
packets. This is now working for observation packets. Besides stop
waiting for replies we needed to send an unobserve packet to
server.

And we now do the same When user calls sol_coap_cancel_send_packet function.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>