Sol-fbp-generator: The generator should provide a context to the metatype.
This will match the behaviour used when the metatype is created by sol-fbp-runner.
A context is given to the metatype with helper functions like: read file and
store type.

The read file function will be usefull when the javascript metatype is
generating its own code, since it needs to parse a javascript file.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>