build: Fix race condition about 'sol_config.h'
When running make even without '-j', FLOW_OIC_GEN rule runs in parallel
with the remaining of Soletta build and was providing 'sol_config.h'
before it was needed most of the time and the build worked. But that
wasn't always the case, which was causing the build to fail for some
users.

This patch adds 'KCONFIG_AUTOHEADER' as dependency for remaining of the
build so it's always there when is needed. This also add a message when
'sol_config.h' is generated to make debug easier.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>