oic/cbor: encode cbor to/from packets on the wire
Don't allocate extra sol_vector sctrucs to hold data from cbor fields
contained in oic packets. Read and write them directly from/to the
packet payload.

To read data from oic packets the user will now get a struct
sol_oic_map_reader, intead of a sol_vector and use the macro
SOL_OIC_MAP_LOOP to loop through all elements contained in the cbor map.

To write data in an oic packet the user won't need to add all elements
in a sol_vector. The user will need to create a callback that receives
a struct sol_oic_map_writer and add all elements to it using
sol_oic_map_append function. The callback strategy was used because the
used doesn't have access to the packet and the cbor element needs to be
created in the packets before adding elements to it and closed after
adding elements.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>