Tests: Be less strict when TEST-OUTPUT family is used to validate tests.
As the commit a38659c0fe3e3efd31408b6b73382ccdc5e1d449, will issue an
warning that /etc/locale.conf is not present in the syste, many tests that
depends on TEST-OUTPUT and TEST-OUTPUT-REGEX will fail.

Since we only want to check if the desired output put is actually there,
the script can be less strict and do not fail the test if more text is
actually printed.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>