log: log domain variables may not be referenced.
At least when the log domain is set to one different than the global
one, if the user issues no log calls in the Soletta app, that variable
will have no use in the code, thus leading to unused variable warnings.
We now tell, then, to the compiler, that the log domain variable may be
unused in the SOL_LOG_INTERNAL_DECLARE() macro.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>