common/mainloop: allow different implementation in runtime.
This moves the mainloop implementation to a structure, allowing it to
be changed in runtime before sol_init() is called.

This makes it easy to integrate Soletta as a guest library with other
applications that already provide a mainloop, such as Node.JS as we
can create a a pass thru wrapper.

It is worth to remember that we can host other guest mainloops, such
as implemented in sol-glib-integration.h.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com