datatypes/vector: +find/match, +insert_at, change insert_sorted return.
Introduce functions to search vectors, such as get the index of the
first or last occurence using linear or binary search.

An alternative is to "match", the pointer may be just a template that
compare_cb() may return 0 to find an element.

Expose insert_at().

Make insert_sorted() return the index so we can skip lookup in some
cases.

Change return types from 'int' to 'int32_t' or 'uint16_t' so we're
sure to work on platforms where int is not 4 bytes

Signed-off-by: Gustavo Sverzut Barbieri <gustavo.barbieri@intel.com>