datatypes/vector: +find/match, +insert_at, change insert_sorted return.
Introduce functions to search vectors, such as get the index of the
first or last occurence using linear or binary search.

An alternative is to "match", the pointer may be just a template that
compare_cb() may return 0 to find an element.

Expose insert_at(), also make it and insert_sorted() return the index
so we can skip lookup in some cases. Actually it would be useful to
change sol_ptr_vector_append(), but there are far more users and I'm
not doing that now -- currently that is easy to get by checking
length.

Signed-off-by: Gustavo Sverzut Barbieri <gustavo.barbieri@intel.com>