common/mainloop: fix possible unsigned integer underflow.
The code was always using "i++" and when no increase was wanted it was
doing "i--" to fix it. But if i == 0, then we have an underflow :-/

Change the code to only increment when we should go to the next
element. If the timeout was re-inserted (r != i), then we must
re-process the current index.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com