modules/keyboard: fix events handling
keyboard module wasn't handling users in the correct place.
When an event happens it should go over list of users
and call the correct on_code function for each user.

The way it was done would call the on_code function of
the node type that registered the watcher first.

With this patch having different keyboard node types
in the same flow will work fine.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>