oic: Perform update on post requests
POST requests can be used to create new resources or to update an
existing resource. As Create is not implemented yet, using same method
used for PUT to update the resource.

OIC reference states that PUT should be used to update the whole content
of a resource and POST should be used to update only a portion of the
resouce. As iotivity is supporting using PUT to update only a portion of
the resource, this was supported in soletta too.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>