oic: Add coap header response code in oic client callbacks
When we receive the response from a request packet or a notification
packet, a callback is called to notify the oic client. Add the coap
response code to this callback, so users can handle error responses.
Also calls the callback even if the payload is empty. This is needed
because some valid responses are empty

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>