coap: Do not remove ACK message if one replies from within a reply callback
coap: Do not remove ack messages from the outgoing list

If one is observing a resource and tries to reply with an ack,
this message will be removed and it will never reach its destination.
In order to be able to send an ack message, sol-coap must not
remove an ack packet from the outgoing list.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>