flow: Add robotics module
This commit adds the robotics module, implementing a simple H-Bridge
node type. Usage is quite simple:

	steering OUT -> IN leftbridge(hbridge)
	emergency_brake OUT -> BRAKE leftbridge
	leftbridge OUT[0] -> IN _(gpio/writer)
	leftbridge OUT[1] -> IN _(gpio/writer)
	leftbridge OUT[2] -> IN _(gpio/writer)
	leftbridge OUT[3] -> IN _(gpio/writer)
	leftbridge THROTTLE -> DUTY_CYCLE_PERENT _(pwm)

The `steering` node will calculate an irange IN value that:

	* If 0, will coast the motor
	* If > 0, will turn the motor clockwise
	* If < 0, will turn the motor counter-clockwise

The THROTTLE output is an integer value in the 0-100 range, and is
calculated by mapping the IN value's value, minimum value, and maximum
value to the desired range.  This is appropriate to connect to a PWM
node's DUTY_CYCLE_PERCENT input port.

The steering node could be, for instance, a skid steering node capable
of obtaining information from quadrature encoders and turn angle and
control two H-Bridges to turn a tank-like robot.

Closes #1261.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>