flow/robotics: Add quadrature encoder node type
Encoders are devices that allow a circuit to know how much something
has moved or turned.  Servo motors employs these encoders, for
instance, but they're commonly used in robotics as separate sensors
where beefier gearboxes are required and commercial servos would be to
expensive or difficult to obtain.

This adds a simple quadrature encoder support, based off of [1].

[1]: https://web.archive.org/web/20150728092206/http://letsmakerobots.com/node/24031

Closes #1262.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>