flow/robotics: Add quadrature encoder node type
Encoders are devices that allow a circuit to know how much something
has moved or turned.  Servo motors employs these encoders, for
instance, but they're commonly used in robotics as separate sensors
where beefier gearboxes are required and commercial servos would be to
expensive or difficult to obtain.

This adds a quadrature encoder node type, whose output is 0 if the
shaft is not spinning, 1 if it's spinning clockwise, and -1 if
counter-clockwise.

Closes #1262.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>