flow/robotics: Implement PID controller
PID (Proportional, Integral, Derivative) controllers are commonly used
in robotics in order to control devices based on sensor inputs and a
set value.  For instance, it is possible to control the temperature of
a furnace by carefully controlling how much heat the heating element
should produce in order to keep the temperature as close to the desired
value as possible.  The implementation is simple (just an expression
with numbers as state), although the tuning process isn't that simple.

Implementation loosely based off of my old implementation in [1].

[1] https://github.com/lpereira/Balancing-Robot/blob/master/Firmware/Firmware.pde#L640

Closes #1263.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>