comms: Don't rely on external definitions
The AF_* definitions will not necessarily be available on all systems,
and even if they are, chances are that by exposing structures that make
use of them, we are forcing applications to depend on OS specific
headers.
Instead, since ou sol_network_addr struct includes a family field,
define the values that we accept there explicitly.

Signed-off-by: Iván Briano <ivan.briano@intel.com>