sol-network: Do not let the user remove the current handle during host_info_cb.
If the user tries to remove a handle that is before the
handle that we're currently iterating, it will mess the
for loop.
As a fix, set the current handle to NULL and if the user tries
to remove it, sol_network_cancel_get_hostname_address_info()
will not be able to find the handle in the list, thus will
retrurn an error.

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>