iio: Add support to read processed channels
Some devices expose IIO channels that are already processed, so it's not
necessary to apply offset and scale on them. Such channels have the
suffix '_input', instead of '_raw' of the raw ones. Now, Soletta IIO
module should be able to read them - it'll first to open the raw one and
them the processed one. In this order because '_raw' ones are way more
common.
This change is internal only and does not affect current API - it should
be transparent to API user which kind of channel is being used.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>