fix dependency-resolver flags check.
This simplifies the code by using Python more efficiently and fix bugs
when the first-try fails.

Always convert cflags and ldflags to lists, then we can easily merge
them with "+" operator, including -Werror prepending.

Add a wrapper to compile_test(), so our code is shorter.

Consider cflags and ldflags independently, callers will take care to
only call with one or another, so the append to variable will be
meaningful, but it should be harmless if both CFLAGS and LDFLAGS are
provided.

Last but not least fix the bug to handle unsupported flags. Before it
was iterating "flags", a string. This was producing char-by-char,
which makes no sense.

More than that, if we try flag by flag with an EMPTY body, -Wextra
will be avoided since it would trigger "argv" and "argc" warnings of
being unused, those are converted to error with -Werror we
prepend. Then we generate warning-free code by casting argc/argv to
void.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com