Soletta modules should link to libsoletta.
If we do not link to libsoletta, then whenever a module (ie: pin-mux,
flow, linux-micro) are loaded from a binary that is NOT linked with
libsoletta, then it will fail to find soletta's symbols.

It didn't show before because all of our tests and binaries are linked
with it. But using it as a node-js module or a simple C application
that calls `dlopen("libsoletta.so")` and calls `sol_init()` may cause
errors.

This should fix #1331.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com