Add pipe2 replacement implementation based on pipe
Systems that don't provide pipe(2) in the first place are unlikely to
need pipe2(2).

This also depends on O_CLOEXEC being defined, which is part of
POSIX.1-2008.  So there's a good chance that the target platforms will
have it. When we run across a platform that is POSIX.1-2001 or earlier,
we'll need to add a suitable #define somewhere.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>