test-fbp: make expected output less strict
Change the regex to look for the content that we explicitly print (the
division by zero), and don't be strict about the other content: the
exact filename and function name that leads to the warning.

These tests currently fail when compiled with Clang due to the different
output it uses for function name when SOL_WRN and friends are used.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>