oic: Scan callback shouldn't set the connected resource
resource->resource variable is used to store which resource was being
observed by this node, if any. This variable was incorrectly set when
scanning.

Before trying to observer a resource that was previously scanned, we
were trying to unobserve it and in some cases this was preventing it to
be observed.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>