oic: Support no payload packets and packets with empty map payload
It was impossible to have a packet with no payload when callback to fill
payload was called. No payload packets were only possible when callback
was NULL.

Add a function to set the type of the payload, so users can create no
payload packets and packets with an empty map as the payload. It is
similar but it is not handled the same way by iotivity and Soletta.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>