flow/robotics: Add skid steering node type
This node type controls the steering of a skid-steered platform, using
as inputs the output of a skid-steering odometer, an angle to turn the
platform, and the desired throttle.

The desired throttle will be clamped from minimum and maximum
throttles, which are options since each robotic platform will have
their own minimum and maximum values.

The throttle output will also be calculated proportional to the error
to the desired angle, with an acceleration and a deceleration ramp
proportional to this error; this ramp should make the in-place turning
slightly quicker should the error be large enough.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>