flow: Don't assume node-type headers are all in the same place
The FBP generator assumes that all node types headers can be found
under sol-flow/ in the include directory. This is wrong, as the
node-type generator may take a 'namespace' parameter where the headers
will be stored.

In order for the generator to do the right thing, store the namespace +
header name in the JSON description generated, so the FBP generator can
make use of it to include the right file.

Signed-off-by: Iván Briano <ivan.briano@intel.com>