common: Take SOL_MAIN_DEFAULT out of sol-mainloop.h
These macros are intended to be used by applications only, so it makes
no sense to have them as part of a header that is included all over
Soletta itself at build time.
It was mostly harmless to keep them there until now, but when
application specific actions need to be taken within the definitions of
these macros, it may introduce problems.

Signed-off-by: Iván Briano <ivan.briano@intel.com>