TMP flow: add sol_flow_single_new() and sample.
TMP: rebase on top of new wallclock that provides ENABLED port, then
TMP: toggle it to show how to send packets.

Sometimes it is desired to use a single nodes, manually feeding
packets and processing those being sent on the node's output port.

It is required to create a new node type for each type to be wrapped
since we must copy the description, port count and get_port_in() and
get_port_out() so the user-visible node provides meaningful
information. However we do expose the inner node (child), similar to
what static-flow does, so one can access it direct and eventually get
more information, like its type, options and so on.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com