comms: lazy netlink and improved failure handling.
instead of always creating the netlink socket and monitoring it, just
do when the network is actually used.

If _on_event() we return 'false', then network->fd will be closed, so
we must null-it. In that case also close the socket and make its
member -1, so we'll reach sol_network_start_netlink() on next attempt
to use it.

This also simplifies the test, it doesn't need to account for for the
netlink socket anymore.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com